enhance verbiage; remove reference to GitHub Release
The GitHub release is a suboptimal way to distribute the extension, since it isn't signed, and thus you can't install it in Beta/Release, and you can only install it in Nightly/Aurora if you've set *xpinstall.signatures.required* to `false`. It's better to distribute the extension via AMO, and we're already doing so anyway, so I've removed the reference to the GitHub Release.